changed int to unsigned long

+++ main.c
@@ -10,8 +10,8 @@
 #include <time.h>
 
 typedef struct node {   //vytvorenie struktury na ukladanie prvkov
-    unsigned int key;   //kluc z funkcie hash
-    unsigned int data;  //data (cislo)
+    unsigned long key;   //kluc z funkcie hash
+    unsigned long data;  //data (cislo)
     struct node *next;  //ukazovatel na dalsi prvok v spajanom zozname
 } NODE;
 
@@ -19,12 +19,12 @@ typedef struct array {  //struktura tabulky, ukazovatel na prvok v danej bunke
     struct node *next;
 } ARRAY;
 
-int resize(int *SIZE);
+unsigned long resize(unsigned long *SIZE);
 
 ARRAY *table;   //globalna premenna table
 
-unsigned int hash(unsigned int num, int SIZE) {     //funkcia hash, dostane cislo a vrati hash kod
-    unsigned int h = 0;
+unsigned long hash(unsigned long num, unsigned long SIZE) {     //funkcia hash, dostane cislo a vrati hash kod
+    unsigned long h = 0;
     do {    //prechadza cifry cisla od zadu  a nasobi ich cislom 31, tieto hodnoty zratava
         h = 31 * h + (num%10);
         num /= 10;
@@ -33,7 +33,7 @@ unsigned int hash(unsigned int num, int SIZE) {     //funkcia hash, dostane cisl
     return h % SIZE;    //nakoniec vrati hodnotu h modulo velkost tabulky
 }
 
-NODE *create_node(unsigned int data, int SIZE) {    //funkcia vytvori uzol spajaneho zoznamu
+NODE *create_node(unsigned long data, unsigned long SIZE) {    //funkcia vytvori uzol spajaneho zoznamu
     NODE *result = NULL;
     result = (NODE*)malloc(sizeof(NODE));
     result->data = data;
@@ -43,9 +43,9 @@ NODE *create_node(unsigned int data, int SIZE) {    //funkcia vytvori uzol spaja
 }
 
 
-void insert(unsigned int num, int *SIZE) {      //funkcia najde miesto kam vlozit prvok a vlozi ho tam
+void insert(unsigned long num, unsigned long *SIZE) {      //funkcia najde miesto kam vlozit prvok a vlozi ho tam
     NODE *toInsert = create_node(num, *SIZE);   //vytvori NODE
-    unsigned int index = toInsert->key;         //najde index (hash kod prvku)
+    unsigned long index = toInsert->key;         //najde index (hash kod prvku)
     NODE *tmp = table[index].next;          //vytvori tmp premennu
     int i = 0;      //v tomto loope funkcia pozera ci sa dany prvok nahodov nema dat na viac ako 6 miesto
     while (tmp != NULL) {   //na jednom indexe, ak ano, zvacsi tabulku
@@ -68,8 +68,8 @@ void insert(unsigned int num, int *SIZE) {      //funkcia najde miesto kam vlozi
 
 }
 
-NODE *search(unsigned int num, int SIZE) { //funkcia hlada prvok v hash tabulke
-    unsigned int index = hash(num, SIZE);   //najde index na ktorom by sa mal prvok nachadzat
+NODE *search(unsigned long num, unsigned long SIZE) { //funkcia hlada prvok v hash tabulke
+    unsigned long index = hash(num, SIZE);   //najde index na ktorom by sa mal prvok nachadzat
     NODE *tmp = table[index].next;      //vytvori tmp premennu
     while (num != tmp->data) {      //v spajanom zozname na danom indexe hlada prvok, ak ho nenajde, vrati NULL
         if(tmp->next == NULL) {
@@ -80,10 +80,10 @@ NODE *search(unsigned int num, int SIZE) { //funkcia hlada prvok v hash tabulke
     return tmp;     //ak ho najde, vrati NODE prvku
 }
 
-void rehash(int *SIZE, int bigger) {    //funkcia rehash je volana z funkcie resize, aktualizuje tabulku,
+void rehash(unsigned long *SIZE, int bigger) {    //funkcia rehash je volana z funkcie resize, aktualizuje tabulku,
                                         // prehashuje vsetky prvky
     NODE *tmp, *tmpb;   //premenne tmp a tmpb (tmpBefore)
-    for (int i = 0; i < *SIZE / bigger; ++i) {  //postupne prechadza vsetky indexy tabulky
+    for (unsigned long i = 0; i < *SIZE / bigger; ++i) {  //postupne prechadza vsetky indexy tabulky
         tmp = table[i].next;    //tmp sa nastavi na prvy prvok na indexe
         while (tmp != NULL) {   //ak tam nieco najde, pokracuje
             if(tmp == table[i].next) {  //ak je to prvy prvok v poradi
@@ -114,7 +114,7 @@ void rehash(int *SIZE, int bigger) {    //funkcia rehash je volana z funkcie res
 }
 
 
-int resize(int *SIZE) { //funkcia resize zvacsi tabulku a zavola funkciu rehash
+unsigned long resize(unsigned long *SIZE) { //funkcia resize zvacsi tabulku a zavola funkciu rehash
     int bigger = 10;    //bigger reprezentuje hodnotu, kolko krat sa ma tabulka zvacsit
     *SIZE *= bigger;    //zvacsi sa premenna SIZE
     ARRAY *tmp;
@@ -125,21 +125,21 @@ int resize(int *SIZE) { //funkcia resize zvacsi tabulku a zavola funkciu rehash
         printf("nepodarilo sa realokovat\n");
         return 0;
     }
-    for (int i = *SIZE / bigger; i < *SIZE; ++i) {  //na novych miestach v tabulke sa nastavia ukazovatele next na NULL
+    for (unsigned long i = *SIZE / bigger; i < *SIZE; ++i) {  //na novych miestach v tabulke sa nastavia ukazovatele next na NULL
         table[i].next = NULL;
     }
     rehash(SIZE, bigger);   //zavola sa funkcia rehash
     return *SIZE;   //vrati sa nova velkost
 }
 
-void print_table(int SIZE) {    //funkcia vypisuje tabulku
-    for (int i = 0; i < SIZE; ++i) {    //prejde vsetky indexy a ak na danom indexe nieco je,
+void print_table(unsigned long SIZE) {    //funkcia vypisuje tabulku
+    for (unsigned long i = 0; i < SIZE; ++i) {    //prejde vsetky indexy a ak na danom indexe nieco je,
                                         //vypise ho aj so spajanym zoznamom v nom
         if(table[i].next != NULL) {
             NODE *tmp = table[i].next;
-            printf("index %.2d -", i);
+            printf("index %.2lu -", i);
             do {
-                printf(" %.2d, ", tmp->data);
+                printf(" %.2lu, ", tmp->data);
                 tmp = tmp->next;
             } while (tmp != NULL);
             printf("\n");
@@ -149,21 +149,21 @@ void print_table(int SIZE) {    //funkcia vypisuje tabulku
 
 int main(void) {
 
-    int size = 100;  //pociatocna velkost tabulky
+    unsigned long size = 100;  //pociatocna velkost tabulky
     table = (ARRAY*)calloc(size, sizeof(ARRAY));    //alokuje sa tabulka
 
-    int max = 1000000;
-    int toFind = 500000;
+    long max = 1000000;
+    long toFind = 500000;
     clock_t start, end;
     double cpu_time_used;
 
     start = clock();    //zapne casovac
 
-    for (int i = 0; i < max; ++i) {     //vytvori tabulku z max cisel (1000000)
+    for (long i = 0; i < max; ++i) {     //vytvori tabulku z max cisel (1000000)
         insert(i, &size);
     }
-    int num;
-    for (int i = 0; i < toFind; ++i) {  //hlada nahodne cisla od 0 po max v tabulke toFind-krat (500000)
+    long num;
+    for (long i = 0; i < toFind; ++i) {  //hlada nahodne cisla od 0 po max v tabulke toFind-krat (500000)
         num = (rand() % (max));
         search(num, size);
     }
